Fixed api, refactored a lot of code

+++ app/Http/Controllers/EditorialContentController.php
@@ -55,247 +55,6 @@ class EditorialContentController extends Controller {
     }
 
     /**
-     * Get Ec info by ID
-     *
-     * @param int $id the Ec id
-     *
-     * @return JsonResponse return the Ec info
-     */
-    public function getEcJson(int $id): JsonResponse {
-        $apiUrl = explode("/", request()->path());
-        try {
-            $model = $this->_getEcModelFromType($apiUrl[2]);
-        } catch (Exception $e) {
-            return response()->json(['code' => 400, 'error' => $e->getMessage()], 400);
-        }
-
-        $ec = $model::find($id);
-        if (is_null($ec)) {
-            return response()->json(['code' => 404, 'error' => "Not Found"], 404);
-        }
-
-        if ('media' !== $apiUrl[2]) {
-            $downloadUrls = [
-                'geojson' => route('api.ec.' . $apiUrl[2] . '.download.geojson', ['id' => $id]),
-                'gpx' => route('api.ec.' . $apiUrl[2] . '.download.gpx', ['id' => $id]),
-                'kml' => route('api.ec.' . $apiUrl[2] . '.download.kml', ['id' => $id]),
-            ];
-            $geojson = $ec->getGeojson($downloadUrls);
-            if ($ec->featureImage) {
-                $geojson['properties']['image'] = json_decode($ec->featureImage->getJson(), true);
-            }
-
-            if ($ec->ecMedia) {
-                $gallery = [];
-                $ecMedia = $ec->ecMedia;
-                foreach ($ecMedia as $media) {
-                    $gallery[] = json_decode($media->getJson(), true);
-                }
-
-                if (count($gallery)) {
-                    $geojson['properties']['imageGallery'] = $gallery;
-                }
-            }
-        } else {
-            $geojson = $ec->getGeojson([]);
-        }
-
-        // Add Taxonomies
-        $taxonomies = $this->_getTaxonomies($ec, $names = ['activity', 'theme', 'where', 'who', 'when']);
-        $geojson['properties']['taxonomy'] = $taxonomies;
-
-        if ('track' == $apiUrl[2]) {
-            $durations = $this->_getDurations($ec, $names = ['hiking', 'cycling']);
-            $geojson['properties']['duration'] = $durations;
-        }
-
-        return response()->json($geojson);
-    }
-
-    /**
-     * Controller for API api/app/elbrus/{app_id}/geojson/ec_poi_{poi_id}.geojson
-     *
-     * @param int $app_id
-     * @param int $poi_id
-     *
-     * @return JsonResponse
-     */
-    public function getElbrusPoiGeojson(int $app_id, int $poi_id): JsonResponse {
-        $app = App::find($app_id);
-        $poi = EcPoi::find($poi_id);
-        if (is_null($app) || is_null($poi)) {
-            return response()->json(['code' => 404, 'error' => 'Not found'], 404);
-        }
-        $geojson = $poi->getGeojson();
-        // MAPPING
-        $geojson['properties']['id'] = 'ec_poi_' . $poi->id;
-        $geojson = $this->_mapGeojsonPropertyForElbrusApi($geojson, 'contact_phone');
-        $geojson = $this->_mapGeojsonPropertyForElbrusApi($geojson, 'contact_email');
-
-        // Add Taxonomies
-        $taxonomies = $this->_getTaxonomies($poi);
-        $geojson['properties']['taxonomy'] = $taxonomies;
-
-        return response()->json($geojson, 200);
-    }
-
-    private function _getTaxonomies($obj, $names = ['activity', 'theme', 'where', 'who', 'when', 'webmapp_category']) {
-        $taxonomies = [];
-        foreach ($names as $name) {
-            switch ($name) {
-                case 'activity':
-                    $terms = $obj->taxonomyActivities()->pluck('id')->toArray();
-                    break;
-                case 'theme':
-                    $terms = $obj->taxonomyThemes()->pluck('id')->toArray();
-                    break;
-                case 'where':
-                    $terms = $obj->taxonomyWheres()->pluck('id')->toArray();
-                    break;
-                case 'who':
-                    $terms = $obj->taxonomyTargets()->pluck('id')->toArray();
-                    break;
-                case 'when':
-                    $terms = $obj->taxonomyWhens()->pluck('id')->toArray();
-                    break;
-                case 'webmapp_category':
-                    $terms = $obj->taxonomyPoiTypes()->pluck('id')->toArray();
-                    break;
-            }
-            if (count($terms) > 0) {
-                foreach ($terms as $term) {
-                    $taxonomies[$name][] = $name . '_' . $term;
-                }
-            }
-        }
-
-        return $taxonomies;
-    }
-
-    private function _getDurations($obj, $names = ['hiking', 'cycling']) {
-        $durations = [];
-        $activityTerms = $obj->taxonomyActivities()->whereIn('identifier', $names)->get()->toArray();
-        if (count($activityTerms) > 0) {
-            foreach ($activityTerms as $term) {
-                $durations[$term['identifier']] = [
-                    'forward' => $term['pivot']['duration_forward'],
-                    'backward' => $term['pivot']['duration_backward'],
-                ];
-            }
-        }
-
-        return $durations;
-    }
-
-    /**
-     * Controller for API api/app/elbrus/{app_id}/geojson/ec_track_{poi_id}.geojson
-     *
-     * @param int $app_id
-     * @param int $poi_id
-     *
-     * @return JsonResponse
-     */
-    public function getElbrusTrackGeojson(int $app_id, int $track_id): JsonResponse {
-        $app = App::find($app_id);
-        $track = EcTrack::find($track_id);
-        if (is_null($app) || is_null($track)) {
-            return response()->json(['code' => 404, 'error' => 'Not found'], 404);
-        }
-
-        return response()->json($this->_getElbrusTracksGeojsonComplete($app_id, $track_id), 200);
-    }
-
-    /**
-     * Controller for API api/app/elbrus/{app_id}/geojson/ec_track_{poi_id}.json
-     *
-     * @param int $app_id
-     * @param int $poi_id
-     *
-     * @return JsonResponse
-     */
-    public function getElbrusTrackJson(int $app_id, int $track_id): JsonResponse {
-        $app = App::find($app_id);
-        $track = EcTrack::find($track_id);
-        if (is_null($app) || is_null($track)) {
-            return response()->json(['code' => 404, 'error' => 'Not found'], 404);
-        }
-        $geojson = $this->_getElbrusTracksGeojsonComplete($app_id, $track_id);
-
-        return response()->json($geojson['properties'], 200);
-    }
-
-    private function _getElbrusTracksGeojsonComplete(int $app_id, int $track_id): array {
-        $app = App::find($app_id);
-        $track = EcTrack::find($track_id);
-        $geojson = $track->getGeojson();
-        // MAPPING COLON
-        $geojson['properties']['id'] = 'ec_track_' . $track->id;
-        $fields = [
-            'ele_from', 'ele_to', 'ele_max', 'ele_min', 'duration_forward', 'duration_backward'
-        ];
-        foreach ($fields as $field) {
-            $geojson = $this->_mapGeojsonPropertyForElbrusApi($geojson, $field);
-        }
-        // Add featureImage
-        if ($track->featureImage) {
-            $geojson['properties']['image'] = json_decode($track->featureImage->getJson(), true);
-        }
-
-        if ($track->ecMedia) {
-            $gallery = [];
-            $ecMedia = $track->ecMedia;
-            foreach ($ecMedia as $media) {
-                $gallery[] = json_decode($media->getJson(), true);
-            }
-
-            if (count($gallery)) {
-                $geojson['properties']['imageGallery'] = $gallery;
-            }
-            $geojson['properties']['gpx_url'] = route('api.ec.track.download.gpx', ['id' => $track_id]);
-            $geojson['properties']['kml_url'] = route('api.ec.track.download.kml', ['id' => $track_id]);
-        }
-
-        if ($track->ecPois) {
-            $related = [];
-            $pois = $track->ecPois;
-            foreach ($pois as $poi) {
-                $related['poi']['related'][] = $poi->id;
-            }
-
-            $geojson['properties']['related'] = $related;
-        }
-
-        // Add Taxonomies
-        $taxonomies = $this->_getTaxonomies($track, $names = ['activity', 'theme', 'where', 'who', 'when']);
-        $geojson['properties']['taxonomy'] = $taxonomies;
-
-        $durations = $this->_getDurations($track, $names = ['hiking', 'cycling']);
-        $geojson['properties']['duration'] = $durations;
-
-        return $geojson;
-    }
-
-    /**
-     * Convert $geojson['properties']['example_nocolon'] to
-     * $geojson['properties']['example:colon']. If parameter $field_with_colon is left null
-     * then is derived from $filed using the rule "_" -> ":"
-     *
-     * @param      $geojson
-     * @param      $field
-     * @param null $field_with_colon
-     */
-    private function _mapGeojsonPropertyForElbrusApi($geojson, $field, $field_with_colon = null) {
-        if (isset($geojson['properties'][$field])) {
-            if (is_null($field_with_colon)) {
-                $field_with_colon = preg_replace('/_/', ':', $field);
-            }
-            $geojson['properties'][$field_with_colon] = $geojson['properties'][$field];
-        }
-
-        return $geojson;
-    }
-
-    /**
      * Get Ec image by ID
      *
      * @param int $id the Ec id
@@ -368,10 +127,13 @@ class EditorialContentController extends Controller {
         $ecMedia->save();
 
         try {
-            $headers = get_headers($request->url);
+            $url = $request->url;
+            if (isset($url) && substr($url, 0, 4) === 'http') {
+                $headers = get_headers($request->url);
 
-            if (stripos($headers[0], "200 OK") >= 0)
-                Storage::disk('public')->delete($actualUrl);
+                if (stripos($headers[0], "200 OK") >= 0)
+                    Storage::disk('public')->delete($actualUrl);
+            }
         } catch (Exception $e) {
             Log::warning($e->getMessage());
         }
@@ -423,15 +185,13 @@ class EditorialContentController extends Controller {
      * @param int     $id
      * @param string  $format
      *
-     * @return Response
+     * @return mixed
      */
     public function downloadEcPoi(Request $request, int $id, string $format = 'geojson') {
         $ecPoi = EcPoi::find($id);
 
-        $response = response()->json(['code' => 404, 'error' => "Not Found"], 404);
-        if (is_null($ecPoi)) {
-            return $response;
-        }
+        if (is_null($ecPoi))
+            return response()->json(['code' => 404, 'error' => "Not Found"], 404);
 
         $headers = [];
         $downloadUrls = [
@@ -466,13 +226,13 @@ class EditorialContentController extends Controller {
     /**
      * Return EcTrack JSON.
      *
-     * @param Request
-     * @param int   $id
-     * @param array $headers
+     * @param Request $request
+     * @param int     $id
+     * @param array   $headers
      *
-     * @return Response.
+     * @return JsonResponse
      */
-    public function viewEcGeojson(Request $request, int $id, array $headers = []) {
+    public function viewEcGeojson(Request $request, int $id, array $headers = []): JsonResponse {
         $apiUrl = explode("/", request()->path());
         try {
             $model = $this->_getEcModelFromType($apiUrl[2]);
@@ -480,30 +240,19 @@ class EditorialContentController extends Controller {
             return response()->json(['code' => 400, 'error' => $e->getMessage()], 400);
         }
 
-        $response = response()->json(['code' => 404, 'error' => "Not Found"], 404);
-
         $ec = $model::find($id);
-        if (is_null($ec)) {
-            return $response;
-        }
+        if (is_null($ec))
+            return response()->json(['code' => 404, 'error' => "Not Found"], 404);;
 
-        $downloadUrls = [
-            'geojson' => route('api.ec.' . $apiUrl[2] . '.download.geojson', ['id' => $id]),
-            'gpx' => route('api.ec.' . $apiUrl[2] . '.download.gpx', ['id' => $id]),
-            'kml' => route('api.ec.' . $apiUrl[2] . '.download.kml', ['id' => $id]),
-        ];
-        $content = $ec->getGeojson($downloadUrls);
-        $response = response()->json($content, 200, $headers);
-
-        return $response;
+        return response()->json($ec->getGeojson(), 200, $headers);
     }
 
     /**
-     * @param Request
-     * @param int   $id
-     * @param array $headers
+     * @param Request $request
+     * @param int     $id
+     * @param array   $headers
      *
-     * @return Response.
+     * @return mixed
      */
     public function viewEcGpx(Request $request, int $id, array $headers = []) {
         $apiUrl = explode("/", request()->path());
@@ -513,25 +262,21 @@ class EditorialContentController extends Controller {
             return response()->json(['code' => 400, 'error' => $e->getMessage()], 400);
         }
 
-        $response = response()->json(['code' => 404, 'error' => "Not Found"], 404);
-
         $ec = $model::find($id);
-        if (is_null($ec)) {
-            return $response;
-        }
+        if (is_null($ec))
+            return response()->json(['code' => 404, 'error' => "Not Found"], 404);
 
         $content = $ec->getGpx();
-        $response = response()->gpx($content, 200, $headers);
 
-        return $response;
+        return response()->gpx($content, 200, $headers);
     }
 
     /**
-     * @param Request
-     * @param int   $id
-     * @param array $headers
+     * @param Request $request
+     * @param int     $id
+     * @param array   $headers
      *
-     * @return Response.
+     * @return mixed
      */
     public function viewEcKml(Request $request, int $id, array $headers = []) {
         $apiUrl = explode("/", request()->path());
@@ -541,26 +286,22 @@ class EditorialContentController extends Controller {
             return response()->json(['code' => 400, 'error' => $e->getMessage()], 400);
         }
 
-        $response = response()->json(['code' => 404, 'error' => "Not Found"], 404);
-
         $ec = $model::find($id);
-        if (is_null($ec)) {
-            return $response;
-        }
+        if (is_null($ec))
+            return response()->json(['code' => 404, 'error' => "Not Found"], 404);
 
         $content = $ec->getKml();
-        $response = response()->kml($content, 200, $headers);
 
-        return $response;
+        return response()->kml($content, 200, $headers);
     }
 
     /**
-     * @param Request
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
-     * @return Response.
+     * @return JsonResponse
      */
-    public function downloadEcGeojson(Request $request, int $id) {
+    public function downloadEcGeojson(Request $request, int $id): JsonResponse {
         $headers['Content-Type'] = 'application/vnd.api+json';
         $headers['Content-Disposition'] = 'attachment; filename="' . $id . '.geojson"';
 
@@ -568,10 +309,10 @@ class EditorialContentController extends Controller {
     }
 
     /**
-     * @param Request
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
-     * @return Response.
+     * @return mixed
      */
     public function downloadEcGpx(Request $request, int $id) {
         $headers['Content-Type'] = 'application/xml';
@@ -581,30 +322,15 @@ class EditorialContentController extends Controller {
     }
 
     /**
-     * @param Request
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
-     * @return Response.
+     * @return mixed
      */
     public function downloadEcKml(Request $request, int $id) {
         $headers['Content-Type'] = 'application/xml';
         $headers['Content-Disposition'] = 'attachment; filename="' . $id . '.kml"';
 
         return $this->viewEcKml($request, $id, $headers);
-    }
-
-    /**
-     * get points near ecTrack in 500 meters
-     *
-     * @param int $idTrack the id of EcTrack
-     *
-     */
-    public function getEcMediaNearTrack(int $idTrack) {
-        $track = EcTrack::find($idTrack);
-
-        $nearQuery = "SELECT ST_Distance(
-        'SRID=4326;POINT(-72.1235 42.3521)'::geometry,
-		'SRID=4326;LINESTRING(-72.1260 42.45, -72.123 42.1546)'::geometry
-	)";
     }
 }
